requirements.txt
# requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Testing==0.8.1
google-generativeai==0.3.0
easyocr==1.7.0
bleach==6.1.0
python-magic==0.4.27
Werkzeug==3.0.0